+++ src/lib.rs
@@ -5,8 +5,8 @@ mod nanonaut;
 mod robot;
 mod bg;
 
-pub const WINDOW_WIDTH: f32 = 800.0;
-pub const WINDOW_HEIGHT: f32 = 600.0;
+pub const WINDOW_WIDTH: u32 = 800;
+pub const WINDOW_HEIGHT: u32 = 600;
 const GROUND_LEVEL: f32 = -250.0;   // center of "ground plane" rect, 100px tall
 
 pub struct AnimsPlugin;
